lat long fixed for events - google api restricting 6/51 rake task calls

+++ db/migrate/20150319231622_add_lat_long_to_events.rb
@@ -1,6 +1,6 @@
 class AddLatLongToEvents < ActiveRecord::Migration
   def change
-  	add_column :events, :latitude, :float
-  	add_column :events, :longitude, :float
+  	add_column :events, :latitude, :decimal
+  	add_column :events, :longitude, :decimal
   end
 end
